requirements.txt
flask 
Pillow
imutils
easyocr
python-opencv
numpy
sqlite3
pandas
pytesseract
plotly
skimage
scikit-image
scikit-learn
nbformat>=4.2.0
ipykernel



